Update Voldemort_PPFT_August_2024
uint16(uint32(0x3C)) == 0x8664: This was likely the problematic part. It checks the PE header for the architecture (0x8664 for x64). The uint32(0x3C) part is supposed to get the offset to the PE header, but the use of uint16 may not work as expected.


uint32(0x3C) gives the offset of the PE header.
Adding + 4 skips over the "PE\0\0" signature to the architecture-specific field, which holds 0x8664 for x64 binaries.